+++ src/style/layers/circle.rs
@@ -20,7 +20,6 @@ impl CircleLayer {
         }
     }
 
-    #[cfg(feature = "json")]
     pub(crate) fn from_ffi_parts(layer_id: String, layer: UniquePtr<layers::CircleLayer>) -> Self {
         Self { layer_id, layer }
     }