+++ src/style/sources/mod.rs
@@ -1,7 +1,11 @@
+mod any;
 mod geojson;
 mod id;
+mod refs;
 mod traits;
 
-pub use geojson::GeoJsonSource;
+pub use any::{AnySource, OpaqueSource};
+pub use geojson::{GeoJsonSource, GeoJsonSourceRefMut};
 pub use id::SourceId;
+pub use refs::{OpaqueSourceRefMut, SourceRefMut, SourceType};
 pub use traits::Source;